Refactor classes into their own files
