+++ lib/signed-xml.js
@@ -4,307 +4,13 @@ const utils = require("./utils");
 const c14n = require("./c14n-canonicalization");
 const execC14n = require("./exclusive-canonicalization");
 const EnvelopedSignature = require("./enveloped-signature").EnvelopedSignature;
-const crypto = require("crypto");
+const hashAlgorithms = require("./hash-algorithms");
+const signatureAlgorithms = require("./signature-algorithms");
 
 /**
- * Hash algorithm implementation
- *
- */
-function SHA1() {
-  this.getHash = function (xml) {
-    const shasum = crypto.createHash("sha1");
-    shasum.update(xml, "utf8");
-    const res = shasum.digest("base64");
-    return res;
-  };
-
-  this.getAlgorithmName = function () {
-    return "http://www.w3.org/2000/09/xmldsig#sha1";
-  };
-}
-
-function SHA256() {
-  this.getHash = function (xml) {
-    const shasum = crypto.createHash("sha256");
-    shasum.update(xml, "utf8");
-    const res = shasum.digest("base64");
-    return res;
-  };
-
-  this.getAlgorithmName = function () {
-    return "http://www.w3.org/2001/04/xmlenc#sha256";
-  };
-}
-
-function SHA512() {
-  this.getHash = function (xml) {
-    const shasum = crypto.createHash("sha512");
-    shasum.update(xml, "utf8");
-    const res = shasum.digest("base64");
-    return res;
-  };
-
-  this.getAlgorithmName = function () {
-    return "http://www.w3.org/2001/04/xmlenc#sha512";
-  };
-}
 
 /**
- * Signature algorithm implementation
- *
- */
-function RSASHA1() {
-  /**
-   * Sign the given string using the given key
-   *
-   */
-  this.getSignature = function (signedInfo, privateKey, callback) {
-    const signer = crypto.createSign("RSA-SHA1");
-    signer.update(signedInfo);
-    const res = signer.sign(privateKey, "base64");
-    if (callback) {
-      callback(null, res);
-    }
-    return res;
-  };
-
-  /**
-   * Verify the given signature of the given string using key
-   *
-   */
-  this.verifySignature = function (str, key, signatureValue, callback) {
-    const verifier = crypto.createVerify("RSA-SHA1");
-    verifier.update(str);
-    const res = verifier.verify(key, signatureValue, "base64");
-    if (callback) {
-      callback(null, res);
-    }
-    return res;
-  };
-
-  this.getAlgorithmName = function () {
-    return "http://www.w3.org/2000/09/xmldsig#rsa-sha1";
-  };
-}
-
-/**
- * Signature algorithm implementation
- *
- */
-function RSASHA256() {
-  /**
-   * Sign the given string using the given key
-   *
-   */
-  this.getSignature = function (signedInfo, privateKey, callback) {
-    const signer = crypto.createSign("RSA-SHA256");
-    signer.update(signedInfo);
-    const res = signer.sign(privateKey, "base64");
-    if (callback) {
-      callback(null, res);
-    }
-    return res;
-  };
-
-  /**
-   * Verify the given signature of the given string using key
-   *
-   */
-  this.verifySignature = function (str, key, signatureValue, callback) {
-    const verifier = crypto.createVerify("RSA-SHA256");
-    verifier.update(str);
-    const res = verifier.verify(key, signatureValue, "base64");
-    if (callback) {
-      callback(null, res);
-    }
-    return res;
-  };
-
-  this.getAlgorithmName = function () {
-    return "http://www.w3.org/2001/04/xmldsig-more#rsa-sha256";
-  };
-}
-
-/**
- * Signature algorithm implementation
- *
- */
-function RSASHA512() {
-  /**
-   * Sign the given string using the given key
-   *
-   */
-  this.getSignature = function (signedInfo, privateKey, callback) {
-    const signer = crypto.createSign("RSA-SHA512");
-    signer.update(signedInfo);
-    const res = signer.sign(privateKey, "base64");
-    if (callback) {
-      callback(null, res);
-    }
-    return res;
-  };
-
-  /**
-   * Verify the given signature of the given string using key
-   *
-   */
-  this.verifySignature = function (str, key, signatureValue, callback) {
-    const verifier = crypto.createVerify("RSA-SHA512");
-    verifier.update(str);
-    const res = verifier.verify(key, signatureValue, "base64");
-    if (callback) {
-      callback(null, res);
-    }
-    return res;
-  };
-
-  this.getAlgorithmName = function () {
-    return "http://www.w3.org/2001/04/xmldsig-more#rsa-sha512";
-  };
-}
-
-function HMACSHA1() {
-  this.verifySignature = function (str, key, signatureValue) {
-    const verifier = crypto.createHmac("SHA1", key);
-    verifier.update(str);
-    const res = verifier.digest("base64");
-    return res === signatureValue;
-  };
-
-  this.getAlgorithmName = function () {
-    return "http://www.w3.org/2000/09/xmldsig#hmac-sha1";
-  };
-
-  this.getSignature = function (signedInfo, privateKey) {
-    const verifier = crypto.createHmac("SHA1", privateKey);
-    verifier.update(signedInfo);
-    const res = verifier.digest("base64");
-    return res;
-  };
-}
-
-function collectAncestorNamespaces(node, nsArray) {
-  if (!nsArray) {
-    nsArray = [];
-  }
-
-  const parent = node.parentNode;
-
-  if (!parent) {
-    return nsArray;
-  }
-
-  if (parent.attributes && parent.attributes.length > 0) {
-    for (let i = 0; i < parent.attributes.length; i++) {
-      const attr = parent.attributes[i];
-      if (attr && attr.nodeName && attr.nodeName.search(/^xmlns:/) !== -1) {
-        nsArray.push({
-          prefix: attr.nodeName.replace(/^xmlns:/, ""),
-          namespaceURI: attr.nodeValue,
-        });
-      }
-    }
-  }
-
-  return collectAncestorNamespaces(parent, nsArray);
-}
-
-/**
- * Extract ancestor namespaces in order to import it to root of document subset
- * which is being canonicalized for non-exclusive c14n.
- *
- * @param {object} doc - Usually a product from `new DOMParser().parseFromString()`
- * @param {string} docSubsetXpath - xpath query to get document subset being canonicalized
- * @param {object} namespaceResolver - xpath namespace resolver
- * @returns {Array} i.e. [{prefix: "saml", namespaceURI: "urn:oasis:names:tc:SAML:2.0:assertion"}]
- */
-function findAncestorNs(doc, docSubsetXpath, namespaceResolver) {
-  const docSubset = xpath.selectWithResolver(docSubsetXpath, doc, namespaceResolver);
-
-  if (!Array.isArray(docSubset) || docSubset.length < 1) {
-    return [];
-  }
-
-  // Remove duplicate on ancestor namespace
-  const ancestorNs = collectAncestorNamespaces(docSubset[0]);
-  const ancestorNsWithoutDuplicate = [];
-  for (let i = 0; i < ancestorNs.length; i++) {
-    let notOnTheList = true;
-    for (const v in ancestorNsWithoutDuplicate) {
-      if (ancestorNsWithoutDuplicate[v].prefix === ancestorNs[i].prefix) {
-        notOnTheList = false;
-        break;
-      }
-    }
-
-    if (notOnTheList) {
-      ancestorNsWithoutDuplicate.push(ancestorNs[i]);
-    }
-  }
-
-  // Remove namespaces which are already declared in the subset with the same prefix
-  const returningNs = [];
-  const subsetAttributes = docSubset[0].attributes;
-  for (let j = 0; j < ancestorNsWithoutDuplicate.length; j++) {
-    let isUnique = true;
-    for (let k = 0; k < subsetAttributes.length; k++) {
-      const nodeName = subsetAttributes[k].nodeName;
-      if (nodeName.search(/^xmlns:/) === -1) {
-        continue;
-      }
-      const prefix = nodeName.replace(/^xmlns:/, "");
-      if (ancestorNsWithoutDuplicate[j].prefix === prefix) {
-        isUnique = false;
-        break;
-      }
-    }
-
-    if (isUnique) {
-      returningNs.push(ancestorNsWithoutDuplicate[j]);
-    }
-  }
-
-  return returningNs;
-}
-
-function validateDigestValue(digest, expectedDigest) {
-  let buffer;
-  let expectedBuffer;
-
-  const majorVersion = /^v(\d+)/.exec(process.version)[1];
-
-  if (+majorVersion >= 6) {
-    buffer = Buffer.from(digest, "base64");
-    expectedBuffer = Buffer.from(expectedDigest, "base64");
-  } else {
-    // Compatibility with Node < 5.10.0
-    buffer = new Buffer(digest, "base64");
-    expectedBuffer = new Buffer(expectedDigest, "base64");
-  }
-
-  if (typeof buffer.equals === "function") {
-    return buffer.equals(expectedBuffer);
-  }
-
-  // Compatibility with Node < 0.11.13
-  if (buffer.length !== expectedBuffer.length) {
-    return false;
-  }
-
-  for (let i = 0; i < buffer.length; i++) {
-    if (buffer[i] !== expectedBuffer[i]) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-/**
- * Xml signature implementation
- *
- * @param {string} idMode. Value of "wssecurity" will create/validate id's with the ws-security namespace
- * @param {object} options. Initial configurations
+ * @typedef { import("../index.d.ts").SignedXml}
  */
 class SignedXml {
   constructor(idMode, options = {}) {
@@ -346,17 +52,17 @@ class SignedXml {
     };
 
     this.HashAlgorithms = {
-      "http://www.w3.org/2000/09/xmldsig#sha1": SHA1,
-      "http://www.w3.org/2001/04/xmlenc#sha256": SHA256,
-      "http://www.w3.org/2001/04/xmlenc#sha512": SHA512,
+      "http://www.w3.org/2000/09/xmldsig#sha1": hashAlgorithms.Sha1,
+      "http://www.w3.org/2001/04/xmlenc#sha256": hashAlgorithms.Sha256,
+      "http://www.w3.org/2001/04/xmlenc#sha512": hashAlgorithms.Sha512,
     };
 
     this.SignatureAlgorithms = {
-      "http://www.w3.org/2000/09/xmldsig#rsa-sha1": RSASHA1,
-      "http://www.w3.org/2001/04/xmldsig-more#rsa-sha256": RSASHA256,
-      "http://www.w3.org/2001/04/xmldsig-more#rsa-sha512": RSASHA512,
+      "http://www.w3.org/2000/09/xmldsig#rsa-sha1": signatureAlgorithms.RsaSha1,
+      "http://www.w3.org/2001/04/xmldsig-more#rsa-sha256": signatureAlgorithms.RsaSha256,
+      "http://www.w3.org/2001/04/xmldsig-more#rsa-sha512": signatureAlgorithms.RsaSha512,
       // Disabled by default due to key confusion concerns.
-      // 'http://www.w3.org/2000/09/xmldsig#hmac-sha1': HMACSHA1
+      // 'http://www.w3.org/2000/09/xmldsig#hmac-sha1': SignatureAlgorithms.HmacSha1
     };
   }
 
@@ -367,7 +73,7 @@ class SignedXml {
    */
   enableHMAC() {
     this.SignatureAlgorithms = {
-      "http://www.w3.org/2000/09/xmldsig#hmac-sha1": HMACSHA1,
+      "http://www.w3.org/2000/09/xmldsig#hmac-sha1": signatureAlgorithms.HmacSha1,
     };
     this.getKeyInfoContent = () => null;
   }
@@ -470,7 +176,7 @@ class SignedXml {
      * Search for ancestor namespaces before canonicalization.
      */
     let ancestorNamespaces = [];
-    ancestorNamespaces = findAncestorNs(doc, "//*[local-name()='SignedInfo']");
+    ancestorNamespaces = utils.findAncestorNs(doc, "//*[local-name()='SignedInfo']");
 
     const c14nOptions = {
       ancestorNamespaces: ancestorNamespaces,
@@ -483,7 +189,7 @@ class SignedXml {
      * Search for ancestor namespaces before canonicalization.
      */
     if (Array.isArray(ref.transforms)) {
-      ref.ancestorNamespaces = findAncestorNs(doc, ref.xpath, this.namespaceResolver);
+      ref.ancestorNamespaces = utils.findAncestorNs(doc, ref.xpath, this.namespaceResolver);
     }
 
     const c14nOptions = {
@@ -600,7 +306,7 @@ class SignedXml {
       const hash = this.findHashAlgorithm(ref.digestAlgorithm);
       const digest = hash.getHash(canonXml);
 
-      if (!validateDigestValue(digest, ref.digestValue)) {
+      if (!utils.validateDigestValue(digest, ref.digestValue)) {
         this.validationErrors.push(
           "invalid signature: for uri " +
             ref.uri +
@@ -1218,5 +924,4 @@ exports.SignedXml = SignedXml;
 
 module.exports = {
   SignedXml,
-  findAncestorNs,
 };
