+++ lib/utils.js
@@ -1,4 +1,4 @@
-const select = require("xpath").select;
+const xpath = require("xpath");
 
 function attrEqualsExplicitly(attr, localName, namespace) {
   return attr.localName == localName && (attr.namespaceURI == namespace || !namespace);
@@ -25,10 +25,10 @@ function findAttr(node, localName, namespace) {
   return null;
 }
 
-function findFirst(doc, xpath) {
-  const nodes = select(xpath, doc);
+function findFirst(doc, path) {
+  const nodes = xpath.select(path, doc);
   if (nodes.length == 0) {
-    throw "could not find xpath " + xpath;
+    throw "could not find xpath " + path;
   }
   return nodes[0];
 }
@@ -124,14 +124,136 @@ function derToPem(der, pemLabel) {
   throw new Error("Unknown DER format.");
 }
 
-exports.findAttr = findAttr;
-exports.findChilds = findChilds;
-exports.encodeSpecialCharactersInAttribute = encodeSpecialCharactersInAttribute;
-exports.encodeSpecialCharactersInText = encodeSpecialCharactersInText;
-exports.findFirst = findFirst;
-exports.EXTRACT_X509_CERTS = EXTRACT_X509_CERTS;
-exports.PEM_FORMAT_REGEX = PEM_FORMAT_REGEX;
-exports.BASE64_REGEX = BASE64_REGEX;
-exports.pemToDer = pemToDer;
-exports.derToPem = derToPem;
-exports.normalizePem = normalizePem;
+function collectAncestorNamespaces(node, nsArray) {
+  if (!nsArray) {
+    nsArray = [];
+  }
+
+  const parent = node.parentNode;
+
+  if (!parent) {
+    return nsArray;
+  }
+
+  if (parent.attributes && parent.attributes.length > 0) {
+    for (let i = 0; i < parent.attributes.length; i++) {
+      const attr = parent.attributes[i];
+      if (attr && attr.nodeName && attr.nodeName.search(/^xmlns:/) !== -1) {
+        nsArray.push({
+          prefix: attr.nodeName.replace(/^xmlns:/, ""),
+          namespaceURI: attr.nodeValue,
+        });
+      }
+    }
+  }
+
+  return collectAncestorNamespaces(parent, nsArray);
+}
+
+/**
+ * Extract ancestor namespaces in order to import it to root of document subset
+ * which is being canonicalized for non-exclusive c14n.
+ *
+ * @param {object} doc - Usually a product from `new DOMParser().parseFromString()`
+ * @param {string} docSubsetXpath - xpath query to get document subset being canonicalized
+ * @param {object} namespaceResolver - xpath namespace resolver
+ * @returns {Array} i.e. [{prefix: "saml", namespaceURI: "urn:oasis:names:tc:SAML:2.0:assertion"}]
+ */
+function findAncestorNs(doc, docSubsetXpath, namespaceResolver) {
+  const docSubset = xpath.selectWithResolver(docSubsetXpath, doc, namespaceResolver);
+
+  if (!Array.isArray(docSubset) || docSubset.length < 1) {
+    return [];
+  }
+
+  // Remove duplicate on ancestor namespace
+  const ancestorNs = collectAncestorNamespaces(docSubset[0]);
+  const ancestorNsWithoutDuplicate = [];
+  for (let i = 0; i < ancestorNs.length; i++) {
+    let notOnTheList = true;
+    for (const v in ancestorNsWithoutDuplicate) {
+      if (ancestorNsWithoutDuplicate[v].prefix === ancestorNs[i].prefix) {
+        notOnTheList = false;
+        break;
+      }
+    }
+
+    if (notOnTheList) {
+      ancestorNsWithoutDuplicate.push(ancestorNs[i]);
+    }
+  }
+
+  // Remove namespaces which are already declared in the subset with the same prefix
+  const returningNs = [];
+  const subsetAttributes = docSubset[0].attributes;
+  for (let j = 0; j < ancestorNsWithoutDuplicate.length; j++) {
+    let isUnique = true;
+    for (let k = 0; k < subsetAttributes.length; k++) {
+      const nodeName = subsetAttributes[k].nodeName;
+      if (nodeName.search(/^xmlns:/) === -1) {
+        continue;
+      }
+      const prefix = nodeName.replace(/^xmlns:/, "");
+      if (ancestorNsWithoutDuplicate[j].prefix === prefix) {
+        isUnique = false;
+        break;
+      }
+    }
+
+    if (isUnique) {
+      returningNs.push(ancestorNsWithoutDuplicate[j]);
+    }
+  }
+
+  return returningNs;
+}
+
+function validateDigestValue(digest, expectedDigest) {
+  let buffer;
+  let expectedBuffer;
+
+  const majorVersion = /^v(\d+)/.exec(process.version)[1];
+
+  if (+majorVersion >= 6) {
+    buffer = Buffer.from(digest, "base64");
+    expectedBuffer = Buffer.from(expectedDigest, "base64");
+  } else {
+    // Compatibility with Node < 5.10.0
+    buffer = new Buffer(digest, "base64");
+    expectedBuffer = new Buffer(expectedDigest, "base64");
+  }
+
+  if (typeof buffer.equals === "function") {
+    return buffer.equals(expectedBuffer);
+  }
+
+  // Compatibility with Node < 0.11.13
+  if (buffer.length !== expectedBuffer.length) {
+    return false;
+  }
+
+  for (let i = 0; i < buffer.length; i++) {
+    if (buffer[i] !== expectedBuffer[i]) {
+      return false;
+    }
+  }
+
+  return true;
+}
+
+module.exports = {
+  findAttr,
+  findChilds,
+  encodeSpecialCharactersInAttribute,
+  encodeSpecialCharactersInText,
+  findFirst,
+  EXTRACT_X509_CERTS,
+  PEM_FORMAT_REGEX,
+  BASE64_REGEX,
+  pemToDer,
+  derToPem,
+  normalizePem,
+  collectAncestorNamespaces,
+  findAncestorNs,
+  validateDigestValue,
+};
